Documentation done and upcoming tasks mostly planned - started coding
Mostly changed network, with some code work, towards message routing - need to also adapt for and incorporate names/labels/tags in the routing and storage of 'contents' - also loads of TODOs in onpath, collectors and base. Hashrouting is not actually that useful, as it is mostly just to do with ONE caching strategy. Re-read the paper - KEEP NOTES IN MIND AND COME BACK TO IT AND NOTES FROM TIME TO TIMEgit add -u!