Finished contentplacement, done labels association to nodes
Changed contentplacement, topology and network, to include all remaining content characteristins when placing content and to associate labels and their counts with corresponding nodes (via Counters - the main parts needed for the base of the feedback and control algorithm, implementing the different strategies, changing the placement, depending on the given data). Now need to work with repos, to: make sure that contents are not maintained where not actually in the repo, keep track of all the other characteristics of content, offload the content appropriately, where needed, and implement the storage services appropriately as well, where needed. This should mostly be sorted in the network.py file, as well, but in the view method. Also, in parallel need to have a look at the topology.py and orchestration.py files (again), to better account for repo placement, association of repo characteristics.