Mostly integrated EDR storage - next: network transmission into storage
Implemented EDR storage as a separate entity, so that now IT DOES NOT REPLACE cache, but rather SUPPLEMENTS it. The idea now is that it should store any incoming AND SERVED data, and this needs to be sorted via working with the network.py file (yes, STILL), but in the NetworkView and NetworkController classes, mainly, so that repos provide the appropriate service and routing, and ultimately could be properly managed. EDR/Service Data Storage and Placement Management is the end goal of the implementation and it will be achieved via the feedback and control algorithm and associated service storage strategies.