Updated services for labels and added all_labels location identification
updated servicing (at least partially), as stated in title, and added some precise node identification, for all-label combination content identification (even though that might need some revision, as the node with the most labels of all the content labels may not actually have the content, and need to account for that). Need to now go into the structure of the repos (and any other relevant entity associated with servicing and status changing of content) and implement content removal and accommodation according to freshness period and shelf life.