Added some base code for the random generation with burstiness
The base code, derived from random bursty traffic is copied from the normal random stationary repo workload. The things needed to be done for this is to establish a workload variable for online times, and create the code, to implement some data/request buffering while offline, for sending a traffic burst, at the same time (potentially to different repos) whenever the device comes back online - after this, need to implement storage data upload via the current request workload system, or even create a new module, like the workload system (accounted for in the orchestrator), for that.