Whole system working, with traces appropriately - done most of live system data insertion via requesters
Made the whole system to work, changing the way data travels through the system and accounting for every hop and storing it wherever valid and useful. Also reduced content and service populations to 500 from 1000 and increased request rate, to be close to the trace one, to @200. The implementation of the bursty request traffic and new implementation, of live data insertion via requesters (receivers) also almost completely implemented and the bursty ones also tested - TODO: implementing the storage data (replication) scope in the workload, the mesh repo topology and some more results extraction implementations, for repos (and probably to improve result extraction interfacing of the simulator later, as well).