Storage overloading problem solved with no memory leaks
Had problems with stopping the storage from overloading, but got it sorted by applying the data scopes and limiting the number of replicas for each content/service identifier (and its processed counterparts) in the system. Further, after, there was a problem with too many events being generated. That issue was resolved by adding the replicas to the message replica numbers checks before the messages were stored, as many events would be created before the messages would actually be stored, rendering simulations inefficient. However, this does pose the added complexity and not-exactly-efficient problem of the system marking messages as stored before they are actually stored and not being able to store other messages, processed present in that specific node in the mean time, due to the service being marked as having those replicas in the system already. TODO: That could be done right at some point later, for completeness and further efficiency!